chore(lint): enable scalafix + fatal warnings and clean up the tree
Wire up sbt-scalafix (RemoveUnused) with semanticdb, add -Wunused:imports and
-Xfatal-warnings so warnings now fail the build. Remove unused imports across
the codebase (via scalafix) and fix the remaining lints surfaced by -Werror:

- value-discard: bind discarded results (group.append in ParquetConverter,
  stmt.execute in PostgreSQLWriter)
- numeric widening: XsdResolver Duration(timeout.toLong)
- dead code: drop unreachable branch in LoadToPostgres
- pure expression: remove orphan Repetition.REPEATED in ParquetSchemaMapper
- deprecation: PostgreSQLIntegrationSpec uses DockerImageName.parse

+++ modules/core/src/main/scala/io/slatr/schema/XsdResolver.scala
@@ -67,7 +67,7 @@ class XsdResolver(config: XsdConfig) extends LazyLogging {
     try {
       val request = basicRequest
         .get(uri"$url")
-        .readTimeout(scala.concurrent.duration.Duration(config.timeout, "seconds"))
+        .readTimeout(scala.concurrent.duration.Duration(config.timeout.toLong, "seconds"))
       
       val response = request.send(backend)
